Move logic for backgroundColor, color, and icon to LogItem component
